requirements.txt
flask>=3.0.0
supabase>=2.0.0
groq>=0.31.0
python-dotenv>=1.0.0
requests>=2.31.0
flask-cors>=4.0.0
fuzzywuzzy>=0.18.0
python-levenshtein>=0.21.0
gunicorn>=21.2.0
httpx<0.28.0
